Added a separate utility function for checking if a compressed file is within a time range
